Optimize image and remove used component
